Fixing the sample table now
