+++ src/components/data-management/ProjectDetails.jsx
@@ -1,13 +1,12 @@
 /* eslint-disable react/jsx-props-no-spreading */
-import React from 'react';
+import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 
 import {
-  Space, Col, Row, Typography,
+  Space, Col, Row, Typography, Button,
 } from 'antd';
 import PropTypes from 'prop-types';
-import SamplesTable from './SamplesTable';
-// import '../../utils/css/data-management.css';
+import SamplesTableOld from './SamplesTableOld';
 import ProjectMenu from './ProjectMenu';
 import {
   updateProject,
@@ -21,12 +20,19 @@ const ProjectDetails = ({ width, height }) => {
   const { activeProjectUuid } = useSelector((state) => state.projects.meta) || false;
   const activeProject = useSelector((state) => state.projects[activeProjectUuid]);
 
+  const [creatingMetadataColumn, setCreatingMetadataColumn] = useState(false);
+
+  console.log('***** ', creatingMetadataColumn);
+
   return (
     <div id='project-details' width={width} height={height}>
       <Space direction='vertical' style={{ width: '100%', padding: '8px 4px' }}>
         <Row style={{ display: 'flex', justifyContent: 'space-between' }}>
           <Title level={3}>{activeProject?.name}</Title>
-          <ProjectMenu />
+          <Space>
+            <Button onClick={() => setCreatingMetadataColumn(true)}>Fake add metadata</Button>
+            <ProjectMenu />
+          </Space>
         </Row>
         <Row>
           {activeProjectUuid ? (
@@ -50,8 +56,10 @@ const ProjectDetails = ({ width, height }) => {
           {' '}
 
         </Row>
-        <SamplesTable
+        <SamplesTableOld
           height={height}
+          creatingMetadataColumn={creatingMetadataColumn}
+          setCreatingMetadataColumn={setCreatingMetadataColumn}
         />
       </Space>
     </div>
