+++ src/components/data-management/UploadDetailsModal.jsx
@@ -1,5 +1,6 @@
 /* eslint-disable import/no-duplicates */
 import React, { useState, useEffect, useRef } from 'react';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import moment from 'moment';
 import {
@@ -8,6 +9,7 @@ import {
 import { UploadOutlined } from '@ant-design/icons';
 import { useDispatch } from 'react-redux';
 import { uploadSingleFile, bundleToFile } from '../../utils/upload/processUpload';
+
 import pushNotificationMessage from '../../utils/pushNotificationMessage';
 import UploadStatus, { messageForStatus } from '../../utils/upload/UploadStatus';
 import downloadSingleFile from '../../utils/data-management/downloadSingleFile';
@@ -19,7 +21,7 @@ const SELECTED_TECH = '10X Chromium';
 const UploadDetailsModal = (props) => {
   const dispatch = useDispatch();
   const {
-    sampleName, visible, onCancel, uploadDetailsModalDataRef, activeProjectUuid,
+    visible, onCancel, uploadDetailsModalDataRef,
   } = props;
   const { fileCategory, sampleUuid } = uploadDetailsModalDataRef.current || false;
   const file = uploadDetailsModalDataRef.current?.file || {};
@@ -30,6 +32,10 @@ const UploadDetailsModal = (props) => {
   const bundleName = bundle?.name;
   const inputFileRef = useRef(null);
   const [replacementFileBundle, setReplacementFileBundle] = useState(null);
+
+  const { activeProjectUuid } = useSelector((state) => state.projects.meta) || false;
+  const samples = useSelector((state) => state.samples);
+  const sampleName = samples[uploadDetailsModalDataRef.current?.sampleUuid]?.name;
 
   useEffect(() => {
     if (replacementFileBundle) {
